LayoutTests/media/track/audio-track-add-remove-expected.txt


Test audio track handling.

EVENT(addtrack)
PASS: Track is in valid state.
END OF TEST
